+++ src/comm_protocol.h
@@ -177,7 +177,7 @@ typedef struct {
 typedef struct {
     uint8_t test_RS485;             /**< Variable for testing RS485 */
     uint8_t test_Sync;              /**< Variable for testing Sync */
-    float32_t test_CAN;             /**< Variable for testing the CAN Bus */
+    uint16_t test_CAN;             /**< Variable for testing the CAN Bus */
     bool test_bool_CAN;             /**< Boolean variable for testing the CAN Bus */
     uint16_t analog_value_measure;  /**< Analog measurement */
     uint8_t id_and_status;          /**< Status information */
@@ -236,7 +236,7 @@ extern uint16_t CAN_Bus_receive;
 extern uint16_t CAN_Bus_receive_ref;
 extern bool CAN_Bus_bool_receive;
 
-extern float32_t reference_value;
+extern uint16_t reference_value;
 
 extern bool print_done;
